validaciones y manejo expections flujo user, audio y favorites

+++ app/services/favorites_service.py
@@ -1,3 +1,4 @@
+from ..databases.db import db
 from ..services.config_service import ConfigService
 from ..repositories.favorites_repository import FavoritesRepository
 
@@ -16,17 +17,42 @@ class FavoriteService:
 
             result.append(favorite_data)
 
-        return result
+        return result, 200
     
     @staticmethod 
-    def add_favorite(user_id, audio_id):
-        favorite = FavoritesRepository.get_favorite(user_id, audio_id)
-        if favorite:
-            raise ValueError("El audio ya se está en la lista de favoitos")
+    def add_favorite(audio_id, data):
+        try:
+            
+            if not data:
+                return {"message": "Los datos proporcionados están vacíos"}, 400
+            
+            if 'user_ID' not in data:
+                return {"message": "No se recibió user_ID"}, 400
+            
+            user_id = data.get("user_ID")
+            favorite = FavoritesRepository.get_favorite(user_id, audio_id)
+            if favorite:
+                return {"message":"El audio ya se encuentra en la lista de favoitos"}, 304
+            
+            new_favorite = FavoritesRepository.add_favorite(user_id, audio_id)
+            db.session.commit()
+            
+            return {"new_favorite": new_favorite.to_dict()}, 200
         
-        return FavoritesRepository.add_favorite(user_id, audio_id)
+        except Exception as e:
+            db.session.rollback()
+            return {"message": f'Ocurrió un error: {str(e)}', "error_type": "Unhandled Exception"}, 500
     
     @staticmethod
-    def delete_favorite(user_id, audio_id):
+    def delete_favorite(audio_id, data):
+        if not data:
+                return {"message": "Los datos proporcionados están vacíos"}, 400
+            
+        if 'user_ID' not in data:
+            return {"message": "No se recibió user_ID"}, 400
+        
+        user_id = data.get("user_ID")
         if not FavoritesRepository.delete_favorite(user_id, audio_id):
-            raise ValueError("El favorite no se encontró")
+            return {"message":"El favorito no se encontró"}, 404
+        
+        return {"message": "Favorito eliminado con éxito"}, 200
